requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
boto==2.38.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
Django==1.8.2
django-appconf==1.0.1
django-braces==1.8.0
django-ckeditor==4.4.8
django-dbbackup==2.0.4
django-debug-toolbar==1.8.0
django-imagefit==0.4
django-lockdown==1.1
django-secure==1.0.1
django-storages==1.1.8
django-tagging==0.3.6
djrill==1.4.0
elasticsearch==1.2.0
enum34==1.1.6
Fabric==1.10.0
factory-boy==2.3.1
first==2.0.1
httplib2==0.9.1
idna==2.6
ipaddress==1.0.18
oauth2==1.5.211
oauthlib==1.0.1
paramiko==2.2.1
Pillow==2.8.2
pip-tools==1.9.0
psycopg2==2.7.3.1
pyasn1==0.3.3
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.4.0
PyNaCl==1.1.2
pysftp==0.2.9
python-memcached==1.53
python-openid==2.2.5
python-social-auth==0.2.12
PyYAML==3.11
raven==5.1.1
requests==2.18.4
requests-oauthlib==0.5.0
six==1.10.0
sqlparse==0.2.3
urllib3==1.22
uWSGI==2.0.15
